Significant rewrite. Existing app did not work well on iOS or Android.
Added detection code to provide error prompts and messages when attempting to run native code plugin on Debug tab, that plugin is not compatible with the file layout that the app resides in when run in APPX. Fixed some issues with whitelists and CSP rules. Condensed the app functions into an "app." object, to methods and properties are more consistent. Revised the way the Emulate tab and third-party plugin alerts were being handled and made the messages more consistent. Added comments throughout and caught more error conditions.